Update to provide an implementation of a attribute that where the data is stored on disk using a hdf file. Currently, this is pig slow for real world usage but will implemented a caching idea and hopefully within the 'real' world that should fix the performance issue.

+++ src/rastergis/RSGISCreateNewAttributeTable.h
@@ -37,6 +37,7 @@
 
 #include "rastergis/RSGISAttributeTable.h"
 #include "rastergis/RSGISAttributeTableMem.h"
+#include "rastergis/RSGISAttributeTableHDF.h"
 
 #include "math/RSGISMathsUtils.h"
 
@@ -63,7 +64,7 @@ namespace rsgis{namespace rastergis{
     public:
         RSGISCreateNewAttributeTable();
         RSGISAttributeTable* createAndPopPixelCount(GDALDataset *clumpsDataset, bool useMemory=true, string outFilePath="")throw(RSGISImageCalcException, RSGISAttributeTableException);
-        unsigned long long calcMaxValue(GDALDataset *dataset)throw(RSGISImageCalcException);
+        size_t calcMaxValue(GDALDataset *dataset)throw(RSGISImageCalcException);
         ~RSGISCreateNewAttributeTable();
     };
